% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-mclapply.R
\name{bg_mclapply}
\alias{bg_mclapply}
\title{Multicore lapply in the background}
\usage{
bg_mclapply(X, FUN, mc.cores = 1, ..., .wait = TRUE, .seed = NULL)
}
\arguments{
\item{X}{A list.}

\item{FUN}{The function to be applied to each element of \code{X}.}

\item{mc.cores}{Passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}.}

\item{...}{Arguments passed to \code{FUN}.}
}
\description{
Multicore lapply in the background
}
\examples{
ans <- bg_mclapply(seq(100), sqrt, mc.cores = 2)

}
