% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-bg.R
\name{bg_mrgsim_d}
\alias{bg_mrgsim_d}
\title{Run mrgsim in the background}
\usage{
bg_mrgsim_d(mod, data, ..., split = 1, output = ".")
}
\arguments{
\item{mod}{a model object}

\item{...}{arguments passed to \code{\link[mrgsolve:::mrgsim]{mrgsolve::::mrgsim()}}}

\item{path}{a directory for saving simulated data}
}
\description{
Run mrgsim in the background
}
