% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{ext_stream}
\alias{ext_stream}
\title{Change the extension on file_stream names}
\usage{
ext_stream(x, ext)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{ext}{The new extension.}
}
\description{
Change the extension on file_stream names
}
\examples{
x <- new_stream(3)
x <- ext_stream(x, "feather")
x[[1]]$file

}
\seealso{
\code{\link[=format_stream]{format_stream()}}, \code{\link[=locate_stream]{locate_stream()}}, \code{\link[=new_stream]{new_stream()}}, \code{\link[=file_stream]{file_stream()}},
\code{\link[=file_set]{file_set()}}
}
