% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{file_set}
\alias{file_set}
\title{Generate a sequence of file objects}
\usage{
file_set(
  n,
  where = NULL,
  prefix = NULL,
  pad = TRUE,
  sep = "-",
  ext = "",
  file_only = FALSE
)
}
\arguments{
\item{n}{The number of file names / objects to create.}

\item{where}{An optional file path.}

\item{prefix}{A character prefix for the file name.}

\item{pad}{If \code{TRUE}, numbers will be padded with zeros.}

\item{sep}{Separator character.}

\item{ext}{A file extension, including the dot.}

\item{file_only}{If \code{TRUE} then only return an vector of file names.}
}
\value{
By default a list length \code{n} of lists length 2; each sublist contains the
integer file number as \code{i} and the file name as \code{file}. If \code{file_only} is
\code{TRUE}, then a vector of file names.
}
\description{
Generate a sequence of file objects
}
\examples{

x <- file_set(3, where = "foo/bar")
length(x)
length(x[[1]])
x[[2]]$i
x[[3]]$file

file_set(2, file_only = TRUE)

}
\seealso{
\code{\link[=setup_locker]{setup_locker()}}
}
