% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-set.R
\name{file_stream}
\alias{file_stream}
\alias{object_stream}
\title{Create a stream of outputs and inputs}
\usage{
file_stream(n, dataset = NULL, format = NULL, ...)

object_stream(objects, dataset = NULL, format = NULL, ...)
}
\arguments{
\item{n}{The number of streams to create.}

\item{dataset}{Passed to \code{\link[=setup_locker]{setup_locker()}}; important to note that the
directory will be unlinked if it exists and is an established locker
directory.}

\item{format}{Passed to \code{\link[=format_stream]{format_stream()}}.}

\item{...}{Passed to \code{\link[=file_set]{file_set()}} (if \code{dataset} not provided) or
\code{\link[=setup_locker]{setup_locker()}} (if \code{dataset} is provided).}

\item{objects}{A list of objects or a vector.}
}
\value{
A list with the following elements: \code{i} the position number; \code{file}
the output file name; \code{x} the input object (for \code{object_stream} only).
}
\description{
By stream we mean a list that pre-specifies the list of output file names,
replicate numbers and possibly input objects for a simulation. Use
\code{file_stream} to only generate file names with indices (see \code{\link[=file_set]{file_set()}}).
Use \code{object_stream} to create a \code{file_stream} that contains an input
data structure. Passing \code{dataset} initiates a call to \code{\link[=setup_locker]{setup_locker()}},
which sets up the output directories.
}
\examples{
file_stream(n = 3)
file_stream(n = 2, dataset = file.path(tempdir(), "foo"))

df <- data.frame(ID = c(1,2,3,4))
data <- chunk_by_id(df, nchunk = 2)
object_stream(objects = data)

}
\seealso{
\code{\link[=setup_locker]{setup_locker()}}, \code{\link[=file_set]{file_set()}}, \code{\link[=internalize_fst]{internalize_fst()}}
}
