% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{format_is_set}
\alias{format_is_set}
\alias{is.stream_format}
\title{Check format status of file set item}
\usage{
format_is_set(x)

is.stream_format(x)
}
\arguments{
\item{x}{An object, usually a \code{file_set_item}.}
}
\value{
Logical indicating if \code{x} inherits from one of the stream format classes. .
}
\description{
This can be used to check if a file set item has been assigned an output
format (e.g. \code{fst}, \code{feather}, \code{parquet}, \code{qs} or \code{rds}). If the check returns
\code{FALSE} it would signal that data should be returned rather than calling
\code{\link[=write_stream]{write_stream()}}.
}
