% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-bg.R
\name{get_fst}
\alias{get_fst}
\title{Get the contents of an fst file set}
\usage{
get_fst(path, .as_list = FALSE)
}
\arguments{
\item{path}{the directory to search}

\item{.as_list}{should the results be returned as a list (\code{TRUE}) or a
tibble (\code{FALSE})}
}
\description{
Get the contents of an fst file set
}
\seealso{
\code{\link[=list_fst]{list_fst()}}, \code{\link[=head_fst]{head_fst()}}
}
