% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-bg.R
\name{head_fst}
\alias{head_fst}
\title{Get the head of an fst file set}
\usage{
head_fst(path, n = 5, i = 1)
}
\arguments{
\item{path}{the directory to search}

\item{n}{number of rows to show}

\item{i}{which output output chunk to show}
}
\description{
Get the head of an fst file set
}
\seealso{
\code{\link[=get_fst]{get_fst()}}, \code{\link[=list_fst]{list_fst()}}
}
