% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{internalize_fst}
\alias{internalize_fst}
\alias{get_fst}
\title{Get the contents of an fst file set}
\usage{
internalize_fst(path, .as_list = FALSE, ...)

get_fst(path, .as_list = FALSE, ...)
}
\arguments{
\item{path}{The directory to search.}

\item{.as_list}{Should the results be returned as a list (\code{TRUE}) or a
tibble (\code{FALSE}).}

\item{...}{Not used.}
}
\description{
Get the contents of an fst file set
}
\seealso{
\code{\link[=list_fst]{list_fst()}}, \code{\link[=head_fst]{head_fst()}}
}
