% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{is.file_set_item}
\alias{is.file_set_item}
\title{Check if an object is a file_set_item}
\usage{
is.file_set_item(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical value indicating if \code{x} has the \code{file_set_item} attribute set..
}
\description{
Check if an object is a file_set_item
}
\examples{
x <- new_stream(2)
is.file_set_item(x[[2]])

}
