% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{is.file_stream}
\alias{is.file_stream}
\title{Check if an object inherits from file_stream}
\usage{
is.file_stream(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical value indicating if \code{x} inherits from \code{file_stream}.
}
\description{
Check if an object inherits from file_stream
}
\examples{
x <- new_stream(2)
is.file_stream(x)

}
