% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{is.locker_stream}
\alias{is.locker_stream}
\title{Check if an object inherits from locker_stream}
\usage{
is.locker_stream(x)
}
\arguments{
\item{x}{An object.}
}
\value{
Logical value indicating if \code{x} inherits from \code{locker_stream}.
}
\description{
Check if an object inherits from locker_stream
}
\examples{
x <- new_stream(2, locker = temp_ds("locker-stream-example"))
is.locker_stream(x)

}
