% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{is_locker_dir}
\alias{is_locker_dir}
\title{Check if a directory is dedicated locker space}
\usage{
is_locker_dir(where)
}
\arguments{
\item{where}{The locker location.}
}
\description{
Check if a directory is dedicated locker space
}
