% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-bg.R
\name{list_fst}
\alias{list_fst}
\title{List all bg output files in a directory}
\usage{
list_fst(path)
}
\arguments{
\item{path}{the directory to search}
}
\description{
List all bg output files in a directory
}
