% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\name{mrgsim.parallel}
\alias{mrgsim.parallel}
\title{Simulate with 'mrgsolve' in Parallel}
\description{
Simulate with 'mrgsolve' in Parallel
}
\section{Resources}{

\itemize{
\item GitHub: \url{https://github.com/kylebaron/mrgsim.parallel}
\item Docs: \url{https://kylebaron.github.io/mrgsim.parallel}
}
}

\section{Package options}{

\itemize{
\item \code{mrgsim.parallel.mc.able}: if \code{TRUE}, multicore will be used if appropriate.
}
}

