% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim_worker.R
\name{mrgsim_ms}
\alias{mrgsim_ms}
\alias{mrgsim_worker}
\title{Run mrgsim after trying to load the shared object}
\usage{
mrgsim_ms(mod, ...)

mrgsim_worker(mod, ...)
}
\arguments{
\item{mod}{a model object}

\item{...}{passed to \code{\link[mrgsolve:mrgsim]{mrgsolve::mrgsim()}}}
}
\description{
Use this function when running mrgsolve while parallelizing on a
multisession worker node where the model dll might not be loaded.
}
\examples{

mrgsim_worker(mrgsolve:::house())

}
