% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parallel.R
\name{parallel_mrgsim_d}
\alias{parallel_mrgsim_d}
\alias{future_mrgsim_d}
\alias{mc_mrgsim_d}
\title{Simulate a data set in parallel}
\usage{
future_mrgsim_d(mod, data, nchunk = 4, ..., as_list = FALSE)

mc_mrgsim_d(mod, data, nchunk = 4, ..., as_list = FALSE)
}
\arguments{
\item{mod}{mrgsolve model object}

\item{data}{data set to simulate}

\item{nchunk}{number of chunks in which to split the data set}

\item{...}{passed to \link{mrgsim_d}}

\item{as_list}{if \code{TRUE} a list is return; otherwise (default) a data frame}
}
\value{
A data frame or list of simulated data
}
\description{
Use \link{future_mrgsim_d} to simulate with the \code{future} package.  Use
\link{mc_mrgsim_d} to simulate with \link[parallel:mclapply]{parallel::mclapply}.
}
