% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_parallel.R
\name{parallel_mrgsim_ei}
\alias{parallel_mrgsim_ei}
\alias{future_mrgsim_ei}
\alias{mc_mrgsim_ei}
\title{Simulate an idata set in parallel}
\usage{
future_mrgsim_ei(mod, event, idata, nchunk = 4, ..., as_list = FALSE)

mc_mrgsim_ei(mod, event, idata, nchunk = 4, ..., as_list = FALSE)
}
\arguments{
\item{mod}{mrgsolve model object}

\item{event}{a mrgsolve event object}

\item{idata}{\code{idata} set to simulate}

\item{nchunk}{number of chunks in which to split the data set}

\item{...}{passed to \link{mrgsim_ei}}

\item{as_list}{if \code{TRUE} a list is return; otherwise (default) a data frame}
}
\value{
A data frame or list of simulated data
}
\description{
Use \link{future_mrgsim_ei} to simulate with the \code{future} package.  Use
\link{mc_mrgsim_ei} to simulate with \link[parallel:mclapply]{parallel::mclapply}.
}
\examples{

mod <- modlib("pk2")

event <- ev(amt = 100)

idata <- data.frame(CL = runif(10, 0.5, 1.5))

out <- future_mrgsim_ei(mod, event, idata)

}
