% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{retire_locker}
\alias{retire_locker}
\title{Retire an existing locker location}
\usage{
retire_locker(where)
}
\arguments{
\item{where}{the locker location}
}
\description{
This function removes the the hidden locker file which designates a directory
as a locker. Once the locker is retired, it cannot be reset again by
calling \code{\link[=setup_locker]{setup_locker()}} or \code{\link[=new_stream]{new_stream()}}.
}
\seealso{
\code{\link[=setup_locker]{setup_locker()}}, \code{\link[=reset_locker]{reset_locker()}}
}
