% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrgsim-bg.R
\name{setup_locker}
\alias{setup_locker}
\title{Set up a data storage locker}
\usage{
setup_locker(
  locker,
  tag = locker_tag(locker),
  n = 0,
  .format = "fst",
  write_dummy = FALSE
)
}
\arguments{
\item{locker}{the directory that contains tagged directories of run
results}

\item{tag}{the name of a folder under \code{locker}; this directory must not
exist the first time the locker is set up and will be deleted and re-created
each time it is used to store output from a new simulation run}

\item{n}{the number of output files to be saved}

\item{.format}{the format extension for output files; use \code{fst} to create
outputs using \code{\link[fst:write_fst]{fst::write_fst()}}; use \code{feather} to create outputs using}

\item{write_dummy}{if \code{TRUE} placeholder files are created containing
tibbles with no rows}
}
\description{
Set up a data storage locker
}
\details{
The \code{arrow} package must be installed to \code{write_dummy} files when \code{.format}
is \code{feather}.
}
\examples{

x <- setup_locker(tempdir(), tag = "my-sims", n = 5)
x

}
