% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{sim_locker}
\alias{sim_locker}
\alias{setup_locker}
\title{Set up a data storage locker}
\usage{
sim_locker(..., file_only = FALSE)

setup_locker(dir, tag = locker_tag(dir), n = 0, ext = ".fst")
}
\arguments{
\item{...}{Arguments passed to \code{setup_locker}.}

\item{file_only}{If \code{TRUE} then only return an vector of file names.}

\item{dir}{The directory that contains tagged directories of run
results.}

\item{tag}{The name of a folder under \code{dir}; this directory must not
exist the first time the locker is set up and \strong{will be deleted} and
re-created each time it is used to store output from a new simulation run.}

\item{n}{The number of output files to be saved.}

\item{ext}{The format extension for output files, including \code{.}.}
}
\value{
When \code{n} is given, \code{setup_locker} returns a character vector of target file
names; \code{sim_locker} returns a list of objects, containing the file name
(\code{file}) and an index of the list position (\code{i}). When \code{n} is not given, an
empty list is returned.
}
\description{
A locker is a directory structure where an enclosing folder contains
subfolders that in turn contain the results of different simulation runs.
When the number of simulation result sets is known, a stream of file names
(\code{setup_locker}) or file name objects (\code{sim_locker}) can be returned, which
can then be fed into any number of parallel apply-like functions.
}
\details{
\code{dir} must exist when setting up the locker. The directory \code{tag} will be
created under \code{dir} and must not exist except if it had previously been
set up using \code{setup_locker}. Established \code{tag} directories will have a
hidden file in them indicating that they are established simulation output
folders. When recreating the \code{tag} directory, it will be unlinked and created
new.
}
\examples{
x <- setup_locker(tempdir(), tag = "my-sims", n = 2)
x

y <- sim_locker(tempdir(), tag = "my-sim-2", n = 2)
y

}
\seealso{
\code{\link[=file_set]{file_set()}}
}
