% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{stream_ext}
\alias{stream_ext}
\title{Change the extension on stream_file names}
\usage{
stream_ext(x, ext)
}
\arguments{
\item{x}{A \code{stream_file} object.}

\item{ext}{The new extension.}
}
\description{
Change the extension on stream_file names
}
\seealso{
\code{\link[=stream_format]{stream_format()}}, \code{\link[=stream_file]{stream_file()}}
}
