% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-set.R
\name{stream_file}
\alias{stream_file}
\title{Create a stream of files}
\usage{
stream_file(n, dataset = NULL, format = NULL, ...)
}
\arguments{
\item{n}{The number of file names to generate.}

\item{dataset}{Passed to \code{\link[=setup_locker]{setup_locker()}}; important to note that the
directory will be unlinked if it exists and is an established locker
directory.}

\item{format}{Passed to \code{\link[=stream_format]{stream_format()}}.}

\item{...}{Passed to \code{\link[=file_set]{file_set()}}.}
}
\description{
Optionally, setup a locker storage space on disk with a specific file
format (e.g. \code{fst} or \code{feather}).
}
\examples{
stream_file(3)

}
