% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{stream_format}
\alias{stream_format}
\title{Set the format for a stream_file object}
\usage{
stream_format(x, type = c("fst", "feather", "qs", "rds"), set_ext = TRUE)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{type}{The file format type; if \code{feather} is chosen, then a check will
be made to ensure the \code{arrow} package is loaded.}

\item{set_ext}{If \code{TRUE}, the existing extension (if it exists) is stripped
and a new extension is added based on the value of \code{type}.}
}
\description{
The format is set on the file objects inside the list so that the file
object can be used to call a write method. See \code{\link[=stream_write]{stream_write()}}.
}
\examples{
fs <- stream_new(2)
fs <- stream_format(fs, "fst")
fs

}
\seealso{
\code{\link[=stream_locate]{stream_locate()}}, \code{\link[=stream_file]{stream_file()}}
}
