% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{stream_locate}
\alias{stream_locate}
\title{Re-sets the directory for stream_file objects}
\usage{
stream_locate(x, where)
}
\arguments{
\item{x}{A \code{stream_file} object.}

\item{where}{The new location.}
}
\description{
Re-sets the directory for stream_file objects
}
\seealso{
\code{\link[=stream_format]{stream_format()}}, \code{\link[=stream_file]{stream_file()}}
}
