% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-set.R
\name{stream_new}
\alias{stream_new}
\alias{stream_new.list}
\alias{stream_new.numeric}
\alias{stream_new.character}
\title{Create a stream of outputs and inputs}
\usage{
stream_new(x, ...)

\method{stream_new}{list}(x, dataset = NULL, format = NULL, ...)

\method{stream_new}{numeric}(x, ...)

\method{stream_new}{character}(x, ...)
}
\arguments{
\item{x}{A list or vector to template the stream; for the \code{numeric} method,
passing a single number will fill \code{x} with a sequence of that length.}

\item{...}{Passed to \code{\link[=file_set]{file_set()}}.}

\item{dataset}{Passed to \code{\link[=setup_locker]{setup_locker()}}; important to note that the
directory will be unlinked if it exists and is an established locker
directory.}

\item{format}{Passed to \code{\link[=stream_format]{stream_format()}}.}
}
\value{
A list with the following elements: \code{i} the position number; \code{file}
the output file name; \code{x} the input object.
}
\description{
By stream we mean a list that pre-specifies the list of output file names,
replicate numbers and possibly input objects for a simulation.  Passing
\code{dataset} initiates a call to \code{\link[=setup_locker]{setup_locker()}}, which sets up the output
directories.
}
\examples{
stream_new(3)
stream_newm(2, dataset = file.path(tempdir(), "foo"))

df <- data.frame(ID = c(1,2,3,4))
data <- chunk_by_id(df, nchunk = 2)
stream_new(data)

}
\seealso{
\code{\link[=setup_locker]{setup_locker()}}, \code{\link[=file_set]{file_set()}}, \code{\link[=internalize_fst]{internalize_fst()}}
}
