% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{stream_write}
\alias{stream_write}
\alias{stream_write.default}
\alias{stream_write.stream_format_fst}
\alias{stream_write.stream_format_feather}
\alias{stream_write.stream_format_qs}
\alias{stream_write.stream_format_rds}
\title{Writer functions for stream_file objects}
\usage{
stream_write(x, ...)

\method{stream_write}{default}(x, data, ...)

\method{stream_write}{stream_format_fst}(x, data, ...)

\method{stream_write}{stream_format_feather}(x, data, ...)

\method{stream_write}{stream_format_qs}(x, data, ...)

\method{stream_write}{stream_format_rds}(x, data, ...)
}
\arguments{
\item{x}{A \code{stream_file} object.}

\item{...}{Not used.}

\item{data}{An object to write.}
}
\description{
This function will write out objects that have been assigned a format
with either \code{\link[=stream_format]{stream_format()}} or the \code{format} argument to \code{\link[=stream_new]{stream_new()}}.
}
\details{
The default method writes to \code{rds} using \code{\link[=saveRDS]{saveRDS()}} and would be invoked
\code{stream_write} is called without setting the \code{format}.
}
\examples{
ds <- temp_ds("example")

fs <- stream_new(2, dataset = ds, format = "fst")

data <- data.frame(x = rnorm(10))

x <- lapply(fs, stream_write, data = data)

list.files(ds)

reset_locker(ds)

fs <- stream_format(fs, "rds")

x <- lapply(fs, stream_write, data = data)

list.files(ds)

}
\seealso{
\code{\link[=stream_format]{stream_format()}}, \code{\link[=stream_file]{stream_file()}}
}
