% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{write_stream}
\alias{write_stream}
\alias{write_stream.default}
\alias{write_stream.stream_format_fst}
\alias{write_stream.stream_format_feather}
\alias{write_stream.stream_format_qs}
\alias{write_stream.stream_format_rds}
\title{Writer functions for stream_file objects}
\usage{
write_stream(x, ...)

\method{write_stream}{default}(x, data, ...)

\method{write_stream}{stream_format_fst}(x, data, ...)

\method{write_stream}{stream_format_feather}(x, data, ...)

\method{write_stream}{stream_format_qs}(x, data, ...)

\method{write_stream}{stream_format_rds}(x, data, ...)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{...}{Not used.}

\item{data}{An object to write.}
}
\description{
This function will write out objects that have been assigned a format
with either \code{\link[=format_stream]{format_stream()}} or the \code{format} argument to \code{\link[=new_stream]{new_stream()}}.
}
\details{
The default method writes to \code{rds} using \code{\link[=saveRDS]{saveRDS()}} and would be invoked
\code{write_stream} is called without setting the \code{format}.
}
\examples{
ds <- temp_ds("example")

fs <- new_stream(2, dataset = ds, format = "fst")

data <- data.frame(x = rnorm(10))

x <- lapply(fs, write_streame, data = data)

list.files(ds)

reset_locker(ds)

fs <- stream_format(fs, "rds")

x <- lapply(fs, stream_write, data = data)

list.files(ds)

}
\seealso{
\code{\link[=format_stream]{format_stream()}}, \code{\link[=file_stream]{file_stream()}}
}
