% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-mclapply.R
\name{bg_mclapply}
\alias{bg_mclapply}
\title{Multicore lapply in the background}
\usage{
bg_mclapply(X, FUN, mc.cores = 1, ..., .wait = TRUE, .seed = NULL)
}
\arguments{
\item{X}{A list.}

\item{FUN}{The function to be applied to each element of \code{X}.}

\item{mc.cores}{Passed to \code{\link[parallel:mclapply]{parallel::mclapply()}}.}

\item{...}{Arguments passed to \code{FUN}.}

\item{.wait}{If \code{FALSE}, the function returns immediately; if \code{TRUE}, then
wait until the background job is finished.}

\item{.seed}{A \code{numeric} value used to set the seed for the simulation;
this is the only way to control the random number generation for your
simulation.}
}
\value{
A list of output data.
}
\description{
Multicore lapply in the background
}
\examples{
ans <- bg_mclapply(seq(10), sqrt, mc.cores = 2)

}
