% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{ext_stream}
\alias{ext_stream}
\title{Change the extension on file_stream names}
\usage{
ext_stream(x, ext)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{ext}{The new extension.}
}
\description{
Change the extension on file_stream names
}
\seealso{
\code{\link[=format_stream]{format_stream()}}, \code{\link[=file_stream]{file_stream()}}
}
