% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-set.R
\name{file_set}
\alias{file_set}
\title{Generate a sequence of file objects}
\usage{
file_set(n, tag = NULL, where = NULL, pad = TRUE, sep = "-", ext = "")
}
\arguments{
\item{n}{the number of file names / objects to create}

\item{tag}{a character prefix for the file name}

\item{where}{an optional file path}

\item{pad}{if \code{TRUE}, numbers will be padded with zeros}

\item{sep}{separator character}

\item{ext}{a file extension, including the dot}
}
\value{
A list length \code{n} of lists length 2; each sublist contains the integer
file number as \code{i} and the file name as \code{file}.
}
\description{
Generate a sequence of file objects
}
\examples{

x <- file_set(3, where = "foo/bar")
length(x)
length(x[[1]])
x[[2]]$i
x[[3]]$file

}
