% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{file_stream}
\alias{file_stream}
\title{Create a stream of files}
\usage{
file_stream(n, locker = NULL, format = NULL, where = NULL, ...)
}
\arguments{
\item{n}{The number of file names to generate.}

\item{format}{Passed to \code{\link[=format_stream]{format_stream()}}.}

\item{...}{Passed to \code{\link[=file_set]{file_set()}}.}
}
\description{
Optionally, setup a locker storage space on disk with a specific file
format (e.g. \code{fst} or \code{feather}).
}
\examples{
stream_file(3)

}
