% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{format_stream}
\alias{format_stream}
\title{Set the format for a file_stream object}
\usage{
format_stream(x, type = c("fst", "feather", "qs", "rds"), set_ext = TRUE)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{type}{The file format type; if \code{feather} is chosen, then a check will
be made to ensure the \code{arrow} package is loaded.}

\item{set_ext}{If \code{TRUE}, the existing extension (if it exists) is stripped
and a new extension is added based on the value of \code{type}.}
}
\description{
Set the format for a file_stream object
}
\seealso{
\code{\link[=locate_stream]{locate_stream()}}, \code{\link[=file_stream]{file_stream()}}, \code{\link[=object_stream]{object_stream()}}
}
