% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{head_fst}
\alias{head_fst}
\title{Get the head of an fst file set}
\usage{
head_fst(path, n = 5, i = 1)
}
\arguments{
\item{path}{The directory to search.}

\item{n}{Number of rows to show.}

\item{i}{Which output output chunk to show.}
}
\description{
Get the head of an fst file set
}
\seealso{
\code{\link[=get_fst]{get_fst()}}, \code{\link[=list_fst]{list_fst()}}
}
