% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{list_fst}
\alias{list_fst}
\title{List all output files in a fst file set}
\usage{
list_fst(path)
}
\arguments{
\item{path}{The (full) directory path to search.}
}
\description{
Use the function to read all of the \code{.fst} files that were saved when
\code{bg_mrgsim_d} was called and \code{.path} was passed along with \code{.format = "fst"}.
}
