% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{locate_stream}
\alias{locate_stream}
\title{Re-sets the directory for file_stream objects}
\usage{
locate_stream(x, where)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{where}{The new location.}
}
\description{
Re-sets the directory for file_stream objects
}
\examples{
x <- new_stream(5)
x <- locate_stream(x, file.path(tempdir(), "foo"))
x[[1]]$file

}
\seealso{
\code{\link[=format_stream]{format_stream()}}, \code{\link[=ext_stream]{ext_stream()}}, \code{\link[=new_stream]{new_stream()}}, \code{\link[=file_stream]{file_stream()}},
\code{\link[=file_set]{file_set()}}
}
