% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{locate_stream}
\alias{locate_stream}
\title{Re-sets the directory for file_stream objects}
\usage{
locate_stream(x, where)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{where}{The new location.}
}
\description{
Re-sets the directory for file_stream objects
}
\seealso{
\code{\link[=format_stream]{format_stream()}}, \code{\link[=file_stream]{file_stream()}}, \code{\link[=object_stream]{object_stream()}}
}
