% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locker.R
\name{noreset_locker}
\alias{noreset_locker}
\title{Prohibit a locker space from being reset}
\usage{
noreset_locker(where)
}
\arguments{
\item{where}{The locker location.}
}
\value{
A logical value indicating if write ability was successfully revoked.
}
\description{
This function removes the the hidden locker file which designates a directory
as a locker. Once the locker is modified this way, it cannot be reset again
by calling \code{\link[=setup_locker]{setup_locker()}} or \code{\link[=new_stream]{new_stream()}}.
}
\seealso{
\code{\link[=setup_locker]{setup_locker()}}, \code{\link[=reset_locker]{reset_locker()}}, \code{\link[=version_locker]{version_locker()}}
}
