% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{stream_write}
\alias{stream_write}
\alias{stream_write.default}
\alias{stream_write.stream_format_fst}
\alias{stream_write.stream_format_feather}
\alias{stream_write.stream_format_qs}
\alias{stream_write.stream_format_rds}
\title{Writer functions for stream_file objects}
\usage{
stream_write(x, ...)

\method{stream_write}{default}(x, data, ...)

\method{stream_write}{stream_format_fst}(x, data, ...)

\method{stream_write}{stream_format_feather}(x, data, ...)

\method{stream_write}{stream_format_qs}(x, data, ...)

\method{stream_write}{stream_format_rds}(x, data, ...)
}
\arguments{
\item{x}{A \code{stream_file} object.}

\item{...}{Not used.}

\item{data}{An object to write.}
}
\description{
Writer functions for stream_file objects
}
\seealso{
\code{\link[=stream_format]{stream_format()}}, \code{\link[=stream_file]{stream_file()}}
}
