% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream-write.R
\name{write_stream}
\alias{write_stream}
\alias{write_stream.stream_format_fst}
\alias{write_stream.stream_format_feather}
\alias{write_stream.stream_format_qs}
\alias{write_stream.stream_format_rds}
\title{Writer functions for file_stream objects}
\usage{
write_stream(x, ...)

\method{write_stream}{stream_format_fst}(x, data, ...)

\method{write_stream}{stream_format_feather}(x, data, ...)

\method{write_stream}{stream_format_qs}(x, data, ...)

\method{write_stream}{stream_format_rds}(x, data, ...)
}
\arguments{
\item{x}{A \code{file_stream} object.}

\item{...}{Not used.}

\item{data}{An object to write.}
}
\description{
Writer functions for file_stream objects
}
\seealso{
\code{\link[=stream_format]{stream_format()}}, \code{\link[=file_stream]{file_stream()}}, \code{\link[=object_stream]{object_stream()}}
}
